chore: add typescript, eslint and prettier to the code
